更新

+++ src/com/example/dao/UploadImgDao.java
@@ -2,6 +2,6 @@ package com.example.dao;
 
 import com.example.entity.UploadImg;
 
-public interface UploadImgDao extends BaseDao<UploadImg, Long> {
+public interface UploadImgDao extends BaseDao<UploadImg, Integer> {
 
 }
