+++ src/com/example/service/UploadImgService.java
@@ -2,6 +2,6 @@ package com.example.service;
 
 import com.example.entity.UploadImg;
 
-public interface UploadImgService extends BaseService<UploadImg, Long> {
+public interface UploadImgService extends BaseService<UploadImg, Integer> {
 
 }
